fix several bugs in training

+++ mindone/transformers/models/qwen2_5_vl/modeling_qwen2_5_vl.py
@@ -1538,11 +1538,8 @@ class Qwen2_5_VLForConditionalGeneration(Qwen2_5_VLPreTrainedModel, GenerationMi
                 mask_expanded = mask_unsqueezed.expand_as(inputs_embeds)
                 image_mask = mask_expanded
 
-                # TODO: remove cast
-                # image_embeds = image_embeds.to(inputs_embeds.dtype)
-                inputs_embeds = (
-                    inputs_embeds.float().masked_scatter(image_mask, image_embeds.float()).to(inputs_embeds.dtype)
-                )
+                image_embeds = image_embeds.to(inputs_embeds.dtype)
+                inputs_embeds = inputs_embeds.masked_scatter(image_mask, image_embeds)
 
             if pixel_values_videos is not None:
                 pixel_values_videos = pixel_values_videos.type(self.visual.dtype)
@@ -1559,11 +1556,8 @@ class Qwen2_5_VLForConditionalGeneration(Qwen2_5_VLPreTrainedModel, GenerationMi
                 mask_expanded = mask_unsqueezed.expand_as(inputs_embeds)
                 video_mask = mask_expanded
 
-                # TODO: remove cast
-                # video_embeds = video_embeds.to(inputs_embeds.dtype)
-                inputs_embeds = (
-                    inputs_embeds.float().masked_scatter(video_mask, video_embeds.float()).to(inputs_embeds.dtype)
-                )
+                video_embeds = video_embeds.to(inputs_embeds.dtype)
+                inputs_embeds = inputs_embeds.masked_scatter(video_mask, video_embeds)
 
         # if we get 4D attention mask we cannot calculate rope deltas anymore. TODO @raushan fixme
         if position_ids is None and (attention_mask is None or attention_mask.ndim == 2):
